♻️ Refactor: Table component
