+++ client/src/ui/Dropdown/custom/AddToPlaylistItem.tsx
@@ -1,19 +1,23 @@
-import DropdownItem from '../../../Dropdown/DropdownItem.tsx';
+import DropdownItem from '../DropdownItem.tsx';
 import { RiAddFill, RiArrowRightSLine } from 'react-icons/ri';
-import SubMenu from '../../../Dropdown/SubMenu/SubMenu.tsx';
+import SubMenu from '../SubMenu/SubMenu.tsx';
 import { useContext } from 'react';
-import { DropdownContext } from '../../../Dropdown/Dropdown.tsx';
-import { useAppDispatch, useAppSelector } from '../../../../app/hooks.ts';
-import { saveSongToPlaylist } from '../../../../features/playlist/playlistThunks.ts';
+import { DropdownContext } from '../Dropdown.tsx';
+import { useAppDispatch, useAppSelector } from '../../../app/hooks.ts';
+import { saveSongToPlaylist } from '../../../features/playlist/playlistThunks.ts';
 
 interface AddToPlaylistItemProps {
   id: string;
 }
 
 const AddToPlaylistItem = ({ id }: AddToPlaylistItemProps) => {
-  const playlists = useAppSelector((state) => state.user?.data?.playlists);
-  const { openSubMenu } = useContext(DropdownContext);
+  const context = useContext(DropdownContext);
+  if (!context) {
+    throw new Error('AddToPlaylistItem should be used within the Dropdown');
+  }
+  const { openSubMenu } = context;
 
+  const playlists = useAppSelector((state) => state.user?.data?.playlists);
   const dispatch = useAppDispatch();
 
   if (!playlists.length || !id) return null;
