--- client/src/ui/Table/Cells/ActionsCell.tsx
@@ -1,94 +1,0 @@
-import styles from './Cells.module.scss';
-import TableCell from '../TableCell.tsx';
-import {
-  RiDeleteBin6Line,
-  RiHeartFill,
-  RiMoreFill,
-  RiShareForwardBoxFill,
-  RiUserHeartLine,
-} from 'react-icons/ri';
-import Dropdown, { DropdownContext } from '../../Dropdown/Dropdown.tsx';
-import DropdownTrigger from '../../Dropdown/DropdownTrigger.tsx';
-import DropdownList from '../../Dropdown/DropdownList.tsx';
-import DropdownItem from '../../Dropdown/DropdownItem.tsx';
-import EditSongDialog from '../../../features/artist/components/EditSongDialog.tsx';
-import useOutsideClick from '../../../hooks/useOutsideClick.tsx';
-import { forwardRef, useContext } from 'react';
-import { useAppDispatch } from '../../../app/hooks.ts';
-import { deleteSong } from '../../../features/artist/artistThunks.ts';
-
-import AddToPlaylistItem from './custom/AddToPlaylistItem.tsx';
-
-// const RefPasser = forwardRef(({ id }) => {
-//   const { closeDropdown } = useContext(DropdownContext);
-//   const { ref, exceptionRef } = useOutsideClick(closeDropdown);
-//   const dispatch = useAppDispatch();
-//
-//   const handleDeleteSong = (id: string) => {
-//     dispatch(deleteSong({ id }));
-//   };
-//
-//   return (
-//     <DropdownList>
-//       <EditSongDialog ref={exceptionRef} id={id} />
-//
-//       <DropdownItem
-//         PreIcon={RiDeleteBin6Line}
-//         underline={true}
-//         onClick={() => handleDeleteSong(id)}
-//       >
-//         Delete
-//       </DropdownItem>
-//
-//       <AddToPlaylistItem id={id} />
-//
-//       <DropdownItem PreIcon={RiHeartFill}>Save to Liked Songs</DropdownItem>
-//       <DropdownItem PreIcon={RiShareForwardBoxFill}>Share</DropdownItem>
-//       <DropdownItem PreIcon={RiUserHeartLine}>Go to Artist</DropdownItem>
-//     </DropdownList>
-//   );
-// });
-
-// TODO: fix outside click
-const RefPasser = forwardRef(({ id }) => {
-  const dispatch = useAppDispatch();
-
-  const handleDeleteSong = (id: string) => {
-    dispatch(deleteSong({ id }));
-  };
-
-  return (
-    <DropdownList>
-      <EditSongDialog id={id} />
-
-      <DropdownItem
-        PreIcon={RiDeleteBin6Line}
-        underline={true}
-        onClick={() => handleDeleteSong(id)}
-      >
-        Delete
-      </DropdownItem>
-
-      <AddToPlaylistItem id={id} />
-
-      <DropdownItem PreIcon={RiHeartFill}>Save to Liked Songs</DropdownItem>
-      <DropdownItem PreIcon={RiShareForwardBoxFill}>Share</DropdownItem>
-      <DropdownItem PreIcon={RiUserHeartLine}>Go to Artist</DropdownItem>
-    </DropdownList>
-  );
-});
-
-const ActionsCell = ({ id }) => {
-  return (
-    <TableCell className={styles.actionsCell} minimize={true}>
-      <Dropdown>
-        <DropdownTrigger>
-          <RiMoreFill />
-        </DropdownTrigger>
-        <RefPasser id={id} />
-      </Dropdown>
-    </TableCell>
-  );
-};
-
-export default ActionsCell;
