--- client/src/ui/Table/Cells/Cells.module.scss
@@ -1,87 +1,0 @@
-// Index
-.indexCell {
-  @apply w-10 text-base text-neutral-400 text-right;
-
-  svg {
-    @apply hidden;
-  }
-}
-
-tr:hover .indexCell {
-  span {
-    @apply hidden;
-  }
-
-  svg {
-    @apply block text-white;
-  }
-}
-
-// Info
-.infoCell {
-  @apply py-2 flex items-center gap-3;
-
-  &Description span {
-    @apply block;
-  }
-
-  &Name {
-    @apply text-base text-white;
-
-    // TODO: remove after adding Link
-    &:hover {
-      @apply underline cursor-pointer;
-    }
-  }
-
-  &NameFull {
-    grid-template-rows: 1/-1;
-  }
-
-  &Artist {
-    @apply text-sm;
-
-    &:hover {
-      @apply underline  cursor-pointer;
-    }
-  }
-
-  img {
-    @apply w-10 rounded;
-  }
-}
-
-tr:hover .infoCellArtist {
-  @apply text-white;
-}
-
-// Like
-.likeCell svg {
-  @apply w-4 cursor-pointer;
-  transition: color 0.3s ease-out;
-}
-
-.likeCell:not(.likeCellHighlighted) svg {
-  @apply invisible;
-
-  &:hover {
-    @apply text-neutral-200;
-  }
-}
-
-tr:hover .likeCell svg {
-  @apply visible;
-}
-
-.likeCellHighlighted {
-  @apply text-green-500;
-}
-
-// Actions
-.actionsCell div > svg {
-  @apply w-4 text-white cursor-pointer scale-125 invisible;
-}
-
-tr:hover .actionsCell svg {
-  @apply visible;
-}
