+++ client/src/ui/Table/custom/SortedTable/Cells/ActionCell.module.scss
@@ -1,0 +1,9 @@
+
+// Actions
+.actionsCell div > svg {
+  @apply w-4 text-white cursor-pointer scale-125 invisible;
+}
+
+tr:hover .actionsCell svg {
+  @apply visible;
+}
