+++ client/src/ui/Table/custom/SortedTable/Cells/IndexCell.module.scss
@@ -1,0 +1,17 @@
+.indexCell {
+  @apply w-10 text-base text-neutral-400 text-right;
+
+  svg {
+    @apply hidden;
+  }
+}
+
+tr:hover .indexCell {
+  span {
+    @apply hidden;
+  }
+
+  svg {
+    @apply block text-white;
+  }
+}
